Refactor: reorganise input package into trie/transport/intent/binding subpackages; rename performedDropAction to performingDropInput
