+++ src/main/java/btm/sword/Sword.java
@@ -37,9 +37,9 @@ import btm.sword.event.listener.SystemListener;
 import btm.sword.event.listener.WorldListener;
 import btm.sword.event.packet.EntityPacketListener;
 import btm.sword.event.packet.MovementListener;
-import btm.sword.input.InputListener;
-import btm.sword.input.ItemInputDispatchTable;
 import btm.sword.input.binding.InputBindingsRegistrar;
+import btm.sword.input.binding.ItemInputDispatchTable;
+import btm.sword.input.transport.InputListener;
 import btm.sword.interaction.CustomInteractionManager;
 import btm.sword.join.MenuSlotGrid;
 import btm.sword.join.ServerJoinArbiter;
