+++ src/main/java/btm/sword/action/core/UmbralBladeAction.java
@@ -16,8 +16,8 @@ import btm.sword.config.Config;
 import btm.sword.entity.base.Combatant;
 import btm.sword.entity.base.SwordEntity;
 import btm.sword.entity.player.SwordPlayer;
-import btm.sword.input.ActivationContext;
 import btm.sword.input.InputType;
+import btm.sword.input.trie.ActivationContext;
 import btm.sword.runtime.scheduler.PredicateRunnablePair;
 import btm.sword.runtime.scheduler.SwordScheduler;
 import btm.sword.runtime.scheduler.TimeArbiter;
