+++ src/main/java/btm/sword/action/throwing/types/ThrownItem.java
@@ -34,7 +34,7 @@ import btm.sword.entity.base.Combatant;
 import btm.sword.entity.base.SwordEntity;
 import btm.sword.entity.player.SwordPlayer;
 import btm.sword.entity.player.ThrowPhase;
-import btm.sword.input.ActivationContext;
+import btm.sword.input.trie.ActivationContext;
 import btm.sword.item.core.ItemUsageManager;
 import btm.sword.item.core.KeyRegistry;
 import btm.sword.runtime.scheduler.PredicateRunnablePair;
