+++ src/main/java/btm/sword/scene/StaticSceneController.java
@@ -9,7 +9,7 @@ import org.bukkit.entity.Player;
 import btm.sword.Sword;
 import btm.sword.entity.player.SwordPlayer;
 import btm.sword.event.packet.MovementListener;
-import btm.sword.input.ActivationContext;
+import btm.sword.input.trie.ActivationContext;
 import btm.sword.runtime.scheduler.SwordScheduler;
 import btm.sword.runtime.scheduler.TimeArbiter;
 import btm.sword.scene.camera.CameraController;
